Move footprint logic
